bugfix: high cpu usage when client disconnects
when client disconnected, the cpu usage was very high. I fixed it by
using recv with MSG_PEEK to get an idea of the size of the message
first. Because when the client disconnects you just receive messages
with size == 0. So theres a workaround to fix that

another thing was that the whole program terminates, when the connection
terminates. that's because it kills the whole pgrp. i fixed that by
giving the forks a new pgrp